rename OBR module to obr
